BoardGameGeek

Documentation for BoardGameGeek.

BoardGameGeek.buddies

BoardGameGeek.gameinfo

BoardGameGeek.gamereviews

BoardGameGeek.userinfo

BoardGameGeek.userreviews

BoardGameGeek.buddies — Method

buddies(name::String)
buddies(user::BGGUser)

Return list of user names of buddies of a user.

BoardGameGeek.gameinfo — Method

gameinfo(id::Integer)

Get game information and summary of reviews.

BoardGameGeek.gamereviews — Method

gamereviews(id::Integer; kwargs...)

Scrape all reviews for a game

Keyword arguments

pagesize::Integer: Number of reviews per page request. Defaults to API maximum 100.

waittime::Real: Wait time between page requests in seconds. Defaults to 2.0f0.

BoardGameGeek.userinfo — Method

userinfo(name::String)

Get user from name string.

BoardGameGeek.userreviews — Method

userreviews(name::String)
userreviews(user::BGGUser)

Return list of all board game reviews a user wrote (ignoring expansions).